refactor(expression_or_float): convert some int to size_t

+++ poincare/include/poincare/expression_or_float.h
@@ -53,7 +53,7 @@ class ExpressionOrFloat {
    */
   PrintFloat::TextLengths writeText(
       std::span<char> buffer,
-      int numberOfSignificantDigits = k_numberOfSignificantDigits,
+      size_t numberOfSignificantDigits = k_numberOfSignificantDigits,
       Preferences::PrintFloatMode floatDisplayMode =
           Preferences::PrintFloatMode::Decimal) const;
